requirements.txt
aiohttp>=3.8.6
aiosignal>=1.3.1
alembic>=1.12.1
antlr4-python3-runtime>=4.9.3
asteroid-filterbanks>=0.4.0
async-timeout>=4.0.3
attrs>=23.1.0
audioread>=3.0.1
blinker>=1.6.3
certifi>=2023.7.22
cffi>=1.16.0
charset-normalizer>=3.3.2
click>=8.1.7
colorama>=0.4.6
coloredlogs>=15.0.1
colorlog>=6.7.0
contourpy>=1.1.1
cycler>=0.12.1
decorator>=5.1.1
docopt>=0.6.2
einops>=0.7.0
filelock>=3.13.1
Flask>=3.0.0
flatbuffers>=23.5.26
fonttools>=4.43.1
frozenlist>=1.4.0
fsspec>=2023.10.0
huggingface-hub>=0.18.0
humanfriendly>=10.0
HyperPyYAML>=1.2.2
idna>=3.4
itsdangerous>=2.1.2
Jinja2>=3.1.2
joblib>=1.3.2
julius>=0.2.7
kiwisolver>=1.4.5
lazy_loader>=0.3
librosa>=0.10.1
lightning>=2.1.0
lightning-utilities>=0.9.0
llvmlite>=0.41.1
Mako>=1.2.4
markdown-it-py>=3.0.0
MarkupSafe>=2.1.3
matplotlib>=3.8.1
mdurl>=0.1.2
more-itertools>=10.1.0
mpmath>=1.3.0
msgpack>=1.0.7
multidict>=6.0.4
networkx>=3.2.1
numba>=0.58.1
numpy>=1.26.1
omegaconf>=2.3.0
onnxruntime>=1.16.1
openai-whisper @ git+https://github.com/openai/whisper.git@b38a1f20f4b23f3f3099af2c3e0ca95627276ddf
optuna>=3.4.0
packaging>=23.2
pandas>=2.1.2
Pillow>=10.1.0
platformdirs>=3.11.0
pooch>=1.8.0
primePy>=1.3
protobuf>=4.24.4
pyannote.audio @ git+https://github.com/pyannote/pyannote-audio@795b92ab265888c58d160f90ae4d91b7bcc6aa2c
pyannote.core>=5.0.0
pyannote.database>=5.0.1
pyannote.metrics>=3.2.1
pyannote.pipeline>=3.0.1
pycparser>=2.21
pydub>=0.25.1
Pygments>=2.16.1
pyparsing>=3.1.1
python-dateutil>=2.8.2
pytorch-lightning>=2.1.0
pytorch-metric-learning>=2.3.0
pytz>=2023.3.post1
PyYAML>=6.0.1
regex>=2023.10.3
requests>=2.31.0
rich>=13.6.0
ruamel.yaml>=0.18.4
ruamel.yaml.clib>=0.2.8
scikit-learn>=1.3.2
scipy>=1.11.3
semver>=3.0.2
sentencepiece>=0.1.99
shellingham>=1.5.4
six>=1.16.0
sortedcontainers>=2.4.0
soundfile>=0.12.1
soxr>=0.3.7
speechbrain>=0.5.15
SQLAlchemy>=2.0.22
sympy>=1.12
tabulate>=0.9.0
tensorboardX>=2.6.2.2
threadpoolctl>=3.2.0
tiktoken>=0.3.3
torch>=2.1.0
torch-audiomentations>=0.11.0
torch-pitch-shift>=1.2.4
torchaudio>=2.1.0
torchmetrics>=1.2.0
torchvision>=0.16.0
tqdm>=4.66.1
typer>=0.9.0
typing_extensions>=4.8.0
tzdata>=2023.3
urllib3>=2.0.7
Werkzeug>=3.0.1
whisper>=1.1.10
yarl>=1.9.2
